refactor: WinningRank의 패키지를 model로 이동
util보다는 model의 역할이 더욱 강한 이유로, WinningRank의 패키지를 변경
