+++ src/test/kotlin/lotto/util/WinningRankTest.kt
@@ -1,5 +1,6 @@
 package lotto.util
 
+import lotto.model.WinningRank
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
